+++ clawgate/__main__.py
@@ -1,5 +1,0 @@
-"""Allow running as: python -m clawgate"""
-
-from clawgate.main import main
-
-main()