+++ foundrygate/__main__.py
@@ -1,0 +1,5 @@
+"""Allow running as: python -m foundrygate"""
+
+from foundrygate.main import main
+
+main()